% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{appendList}
\alias{appendList}
\title{appendList}
\usage{
appendList(l, x, name = NULL)
}
\arguments{
\item{l}{list}

\item{x}{element}

\item{name}{name in list}
}
\value{
list
}
\description{
Function to append an element to a list
}
\examples{
appendList(l = your_list, x = element, name = "TUD")
}
