% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_to_null.R
\name{auto_to_null}
\alias{auto_to_null}
\title{Function that given a param_list iterates tróugh the list and sets every entry to NULL if "auto" was specified}
\usage{
auto_to_null(param_list)
}
\arguments{
\item{param_list}{param_list}
}
\value{
param_list
}
\description{
Function that given a param_list iterates tróugh the list and sets every entry to NULL if "auto" was specified
}
