% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_auto_tuner.R
\name{create_auto_tuner}
\alias{create_auto_tuner}
\title{Function that given a list of graph learners, an inner resampling, a measure, a terminator and a tuner creates a list of auto tuner learners}
\usage{
create_auto_tuner(learner, inner_resampling, measure, terminator, tuner)
}
\arguments{
\item{inner_resampling}{inner resampling strategy}

\item{measure}{measure}

\item{terminator}{termination criterion}

\item{tuner}{tuning procedure#'}

\item{learners}{list of graph_learners}
}
\value{
an mlr3 graph learner
}
\description{
Function that given a list of graph learners, an inner resampling, a measure, a terminator and a tuner creates a list of auto tuner learners
}
\examples{

}
