% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bagging_ensemble.R
\name{create_bagging_ensemble}
\alias{create_bagging_ensemble}
\title{Function that given a list of learners and a task creates an ensemble learner using a bagging approack}
\usage{
create_bagging_ensemble(learners, task_type)
}
\arguments{
\item{learners}{learners}

\item{task}{tesk}
}
\value{
bagging_ensemble
}
\description{
Function that given a list of learners and a task creates an ensemble learner using a bagging approack
}
