% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clones.R
\name{create_clones}
\alias{create_clones}
\title{Function that given a list of learners, creates a list of cloned learners}
\usage{
create_clones(learners)
}
\arguments{
\item{learners}{learners}
}
\value{
list of cloned learners
}
\description{
Function that given a list of learners, creates a list of cloned learners
}
