% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_filter.R
\name{create_feature_filter}
\alias{create_feature_filter}
\title{Function that adds a feature filter to the learner}
\usage{
create_feature_filter(task, type, learners, filter)
}
\arguments{
\item{task}{task}

\item{type}{type}

\item{learners}{list of learners}

\item{filter}{filter}
}
\value{
an mlr3 graph learner
}
\description{
Function that adds a feature filter to the learner
}
\examples{

}
