% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph_learners.R
\name{create_graph_learners}
\alias{create_graph_learners}
\title{Function that given a list of learners and a task creates suitable graphlearners}
\usage{
create_graph_learners(task, learners)
}
\arguments{
\item{task}{task}

\item{learners}{list of learners}
}
\value{
an mlr3 graph learner
}
\description{
Function that given a list of learners and a task creates suitable graphlearners
}
\examples{

}
