% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_learners.R
\name{create_learners}
\alias{create_learners}
\title{Function that given "classif" or "regr" returns a suitable list of learners}
\usage{
create_learners(task, vec_learners = NULL)
}
\arguments{
\item{task}{mlr3 task}

\item{vec_learners}{vector of choosen learners}
}
\value{
list of mlr3 learners
}
\description{
Function that given "classif" or "regr" returns a suitable list of learners
}
\examples{
create_learners(tsk("iris"))
create_learners(tsk("bike_sharing"))

}
