% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_measure.R
\name{create_measure}
\alias{create_measure}
\title{Function that given "classif" or "regr" returns a suitable measure}
\usage{
create_measure(task, measure = NULL)
}
\arguments{
\item{task}{mlr3 task}

\item{measure}{user defined measure}
}
\value{
mlr3 measure
}
\description{
Function that given "classif" or "regr" returns a suitable measure
}
\examples{
create_measure(tsk("iris"))
create_measure(tsk("german_credit"))
create_measure(tsk("bike_sharing"))

}
