% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resampling.R
\name{create_resampling}
\alias{create_resampling}
\title{Function that outputs a MLR resampling object, given specification}
\usage{
create_resampling(task, strat = NULL, params = NULL)
}
\arguments{
\item{task}{mlr3 task}

\item{strat}{resampling strategy}

\item{params}{list of parameters}
}
\value{
resampling object
}
\description{
Function that outputs a MLR resampling object, given specification
}
