% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_robust_learners.R
\name{create_robust_learners}
\alias{create_robust_learners}
\title{Function that given a list of learners and a task creates a robust graphlearners}
\usage{
create_robust_learners(task, learners)
}
\arguments{
\item{task}{task}

\item{learners}{list of learners}
}
\value{
an mlr3 graph learner
}
\description{
Function that given a list of learners and a task creates a robust graphlearners
}
