% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_search_space.R
\name{create_search_space}
\alias{create_search_space}
\title{Function that given a list of learners, defines suitable search_space for HPO}
\usage{
create_search_space(task, learners)
}
\arguments{
\item{task}{taskk}

\item{learners}{list of learners}
}
\value{
list of mlr3 learners with parameter search_space
}
\description{
Function that given a list of learners, defines suitable search_space for HPO
}
