% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stacking_ensemble.R
\name{create_stacking_ensemble}
\alias{create_stacking_ensemble}
\title{Function that given a list of learners and a task and further information creates an ensemble random forrest learner using a stacking approack}
\usage{
create_stacking_ensemble(
  task,
  type,
  learners,
  feature_filter,
  inner_resampling,
  measure,
  terminator,
  tuner
)
}
\arguments{
\item{task}{tesk}

\item{type}{type}

\item{learners}{learners}

\item{feature_filter}{feature_filter}

\item{inner_resampling}{inner resampling}

\item{measure}{measure}

\item{terminator}{terminator}

\item{tuner}{tuner}
}
\value{
stacking_ensemble
}
\description{
Function that given a list of learners and a task and further information creates an ensemble random forrest learner using a stacking approack
}
