% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task.R
\name{create_task}
\alias{create_task}
\title{Function that given a dataframe and variable creates an mlr3 task}
\usage{
create_task(
  df,
  target_var,
  type = NULL,
  rm_na_target = TRUE,
  task_name = "task"
)
}
\arguments{
\item{df}{input data frame}

\item{target_var}{target variable}

\item{type}{type of task. ´regr´ or ´classif´. if ´NULL´ the type is estimated.}

\item{rm_na_target}{remove NAs from target column?}

\item{task_name}{task name}
}
\value{
returns a mlr3 task object
}
\description{
Function that given a dataframe and variable creates an mlr3 task
}
\examples{
create_task(iris, "Species")
create_task(mtcars, 1)

}
