% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_terminator.R
\name{create_terminator}
\alias{create_terminator}
\title{Function that given a task returns a suitable list of learners}
\usage{
create_terminator(vec_terminators = NULL)
}
\arguments{
\item{vec_terminators}{list of termination criteria}
}
\value{
list of termination
}
\description{
Function that given a task returns a suitable list of learners
}
\examples{
create_terminator()

vec_terminators <- list(c("eval", 10), c("rt", 1000))
create_terminator(vec_terminators = vec_terminators)

}
