% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tuner.R
\name{create_tuner}
\alias{create_tuner}
\title{Function that given a tuner returns an mlr tuner object}
\usage{
create_tuner(tuning_method = NULL)
}
\arguments{
\item{tuning_method}{tuner}
}
\value{
tuning method
}
\description{
Function that given a tuner returns an mlr tuner object
}
\examples{
create_tuner()
create_tuner(tuning_method = "cmaes")

}
