% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{debug_console}
\alias{debug_console}
\title{Function to print some logs into the console}
\usage{
debug_console(message, lvl = "INFO")
}
\arguments{
\item{message}{message print to console}

\item{lvl}{level for the message}
}
\description{
Function to print some logs into the console
}
