% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{factor_col_names}
\alias{factor_col_names}
\title{Function to get all column names which are factors}
\usage{
factor_col_names(x)
}
\arguments{
\item{x}{data}
}
\value{
vector of factor columns
}
\description{
Function to get all column names which are factors
}
