% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{feature_importance}
\alias{feature_importance}
\title{Function to get the importance of the features of a mlr3 task}
\usage{
feature_importance(task, filters, ranks = TRUE)
}
\arguments{
\item{task}{mlr task}

\item{filters}{list of lists with filters. list("regr"=list(), "classif" = list())}

\item{ranks}{should ranks be computed. -> make it more robust}
}
\value{
dataframe
}
\description{
Function to get the importance of the features of a mlr3 task
}
