% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{generate_descr_report_tbl_stat}
\alias{generate_descr_report_tbl_stat}
\title{Function for creating statistical text part of descriptive report}
\usage{
generate_descr_report_tbl_stat(data, feature)
}
\arguments{
\item{data}{your data}

\item{feature}{feeature name}
}
\value{
data.frame
}
\description{
Function for creating statistical text part of descriptive report
}
\examples{
generate_descr_report_tbl_stat(data = iris, feature = "Sepal.Length")
}
