% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{generate_descr_report_text_na}
\alias{generate_descr_report_text_na}
\title{Function for creating na text part of descriptive report}
\usage{
generate_descr_report_text_na(
  feature,
  imp_tbl,
  task_obj,
  na_threshold = NULL,
  imp_threshold = NULL
)
}
\arguments{
\item{feature}{name of the feature}

\item{imp_tbl}{table of the calculated importance}

\item{task_obj}{current task object}

\item{na_threshold}{to control the threshold for when a feature has many NAs. If NULL it uses app settings.}

\item{imp_threshold}{to control the threshold for when a feature is important. If NULL it uses app settings.}
}
\value{
returns string
}
\description{
Function for creating na text part of descriptive report
}
