% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{generate_text}
\alias{generate_text}
\title{Function to generate the text with sprintf}
\usage{
generate_text(base_text, text_args)
}
\arguments{
\item{base_text}{text from the corresponding dict}

\item{text_args}{list of args to replace the placeholders}
}
\value{
string
}
\description{
Function to generate the text with sprintf
}
