% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_CR.R
\name{identify_CR}
\alias{identify_CR}
\title{Function to find out, if the Task is Classification or Regression}
\usage{
identify_CR(df, target_var)
}
\arguments{
\item{df}{input data frame}

\item{target_var}{target variable}
}
\value{
returns a string of `regr` or `classif`
}
\description{
Function to find out, if the Task is Classification or Regression
}
\examples{
identify_CR(mtcars, "mpg")
identify_CR(iris, 5)

}
