% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{insert_line_break}
\alias{insert_line_break}
\title{Function insert a line break into a string after n chars}
\usage{
insert_line_break(x, n = 30)
}
\arguments{
\item{x}{string}

\item{n}{insert line break after n chars}
}
\value{
new string with line breaks #377EB8, #E41A1C
}
\description{
Function insert a line break into a string after n chars
}
\examples{
insert_line_break(x = "very long string", n = 10)

}
