% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{MissComb}
\alias{MissComb}
\title{R6 Class for missing combinations
The class has the field for the combinations and custom plot function}
\description{
R6 Class for missing combinations
The class has the field for the combinations and custom plot function

R6 Class for missing combinations
The class has the field for the combinations and custom plot function
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{na_combinations}}{dataframe with missing combinations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MissComb$new()}}
\item \href{#method-plot}{\code{MissComb$plot()}}
\item \href{#method-clone}{\code{MissComb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new MissComb Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MissComb$new(na_comb = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{na_comb}}{dataframe with missing combinations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
custom plot function for the Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MissComb$plot(
  labels = "index",
  label_length = 30,
  show_numbers = TRUE,
  text_offset = 50,
  top_n = 10,
  bar_color = "#FC8D62",
  plot_title = "NA combinations",
  x_lab = "combination",
  y_lab = "n"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{which labels should be used. options: be `index` or `name`. `index` results in a much shorter length usually}

\item{\code{label_length}}{after how many characters the line should be splitted}

\item{\code{show_numbers}}{if `TRUE` shows the numbers of observations next to the bar.}

\item{\code{text_offset}}{defines the offset of the text to the bar}

\item{\code{top_n}}{show the `top_n` cases}

\item{\code{bar_color}}{defines the color of the bars}

\item{\code{plot_title}}{title of the plot}

\item{\code{x_lab}}{x axis label}

\item{\code{y_lab}}{y axis label}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MissComb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
