% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{missing_combinations}
\alias{missing_combinations}
\title{Function to get all combinations of NAs within the rows of a dataset}
\usage{
missing_combinations(x, names_col = TRUE)
}
\arguments{
\item{x}{data}

\item{names_col}{append column with the corresponding column names of the missing combinations}
}
\value{
`MissComb` object
}
\description{
Function to get all combinations of NAs within the rows of a dataset
}
\examples{
missing_combinations(iris, names_col = FALSE)

}
