% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omit_task.R
\name{na_omit_task}
\alias{na_omit_task}
\title{Function that given a task and an na option "omit" overwrites that task with a new task of which the backends na's are omitted}
\usage{
na_omit_task(task, na_option)
}
\arguments{
\item{task}{task}

\item{na_option}{na_option}
}
\value{
task
}
\description{
Function that given a task and an na option "omit" overwrites that task with a new task of which the backends na's are omitted
}
\examples{

}
