% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_auto_ml.R
\name{perform_auto_ml}
\alias{perform_auto_ml}
\title{Function that applies Auto ML Pipeline given a list of parameter}
\usage{
perform_auto_ml(param_list)
}
\arguments{
\item{param_list}{param_list}
}
\value{
benchmark results and measure
}
\description{
Function that applies Auto ML Pipeline given a list of parameter
}
