% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{render_error}
\alias{render_error}
\title{Function to generate an error string}
\usage{
render_error(err_name, cond = NULL, error_list = list_error_mess)
}
\arguments{
\item{err_name}{name of the error within the list}

\item{cond}{passed Condition}

\item{error_list}{list with the errors}
}
\value{
string
}
\description{
Function to generate an error string
}
