% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{ReportContent}
\alias{ReportContent}
\title{R6 Class for missing combinations
The class has the field for the combinations and custom plot function}
\description{
R6 Class for missing combinations
The class has the field for the combinations and custom plot function

R6 Class for missing combinations
The class has the field for the combinations and custom plot function
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{id of object}

\item{\code{type}}{of content currently: text, ggplot, cor_matrix}

\item{\code{content}}{content}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportContent$new()}}
\item \href{#method-print_report}{\code{ReportContent$print_report()}}
\item \href{#method-clone}{\code{ReportContent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new ReportContent Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportContent$new(id = NULL, type = NULL, content = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id of object}

\item{\code{type}}{type of content. currently: text, ggplot, cor_matrix, dataframe}

\item{\code{content}}{content}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_report"></a>}}
\if{latex}{\out{\hypertarget{method-print_report}{}}}
\subsection{Method \code{print_report()}}{
prints the object for the different types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportContent$print_report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportContent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
