% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{ReportPlot}
\alias{ReportPlot}
\title{R6 Class for missing combinations
The class has the field for the combinations and custom plot function}
\description{
R6 Class for missing combinations
The class has the field for the combinations and custom plot function

R6 Class for missing combinations
The class has the field for the combinations and custom plot function
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{dataframe with missing combinations}

\item{\code{descr_key}}{dataframe with missing combinations}

\item{\code{plot}}{dataframe with missing combinations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportPlot$new()}}
\item \href{#method-get_description}{\code{ReportPlot$get_description()}}
\item \href{#method-clone}{\code{ReportPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new ReportPlot Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportPlot$new(title = NULL, descr_key = NULL, descr_dict = NULL, plot = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title of the plot}

\item{\code{descr_key}}{description of the plot}

\item{\code{plot}}{plot to display}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-get_description}{}}}
\subsection{Method \code{get_description()}}{
generates the description for the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportPlot$get_description()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{descr_key}}{key of the description in the dict}

\item{\code{descr_dict}}{dict of description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
