% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_id.R
\name{shorten_id}
\alias{shorten_id}
\title{Function that given a list of learners shortens their id to exclude everything before "regr." or "classif." depending on task type}
\usage{
shorten_id(learners, task_type)
}
\arguments{
\item{learners}{list of learners}

\item{task_type}{task type}
}
\value{
learners with shortened ids
}
\description{
Function that given a list of learners shortens their id to exclude everything before "regr." or "classif." depending on task type
}
\examples{

}
