% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{update_factor_cols}
\alias{update_factor_cols}
\title{Function for updating column}
\usage{
update_factor_cols(data, old_cols, new_cols)
}
\arguments{
\item{data}{your data}

\item{old_cols}{vector of names of old factor columns}

\item{new_cols}{vector of names of new factor columns}
}
\value{
list with new data and new factor columns
}
\description{
Function for updating column
}
