% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_format.R
\name{var_format}
\alias{var_format}
\title{Function that given a vector adjusts formats the datatype in cases that it makes sense.}
\usage{
var_format(vect, n)
}
\arguments{
\item{vect}{vector of data}

\item{n}{number of unique values under below which formating will apply}
}
\value{
formated vector
}
\description{
Function that given a vector adjusts formats the datatype in cases that it makes sense.
}
\examples{
var_format(iris$Species, n = 6)
var_format(mtcars$gear, n = 6)

}
