% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_learners.R
\name{create_learners}
\alias{create_learners}
\title{Function that given a task returns a suitable list of learners}
\usage{
create_learners(task, vec_learners = NULL, hpo = TRUE)
}
\arguments{
\item{task}{mlr3 task}

\item{vec_learners}{vector of choosen learners}

\item{hpo}{choose if searchspaces for hyperparameter optimization should be initialized}
}
\value{
list of mlr3 learners
}
\description{
Function that given a task returns a suitable list of learners
}
\examples{
create_learners(tsk("iris"))
create_learners(tsk("bike_sharing"))
create_learners(task = tsk("german_credit"), vec_learners = c("rf"), hpo = FALSE)

}
