% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resampling.R
\name{create_resampling}
\alias{create_resampling}
\title{Function that outputs a MLR resampling object, igven specification}
\usage{
create_resampling(strat, params)
}
\arguments{
\item{strat}{resampling strategy}

\item{params}{list of parameters}
}
\value{
resampling object
}
\description{
Function that outputs a MLR resampling object, igven specification
}
\examples{
holdout <- list(ration = 0.8)
create_resampling("holdout", holdout)

cv <- list(folds = 2)
create_resampling("cv", cv)

repeated_cv <- list(folds = 2, repeats = 10)
create_resampling("repeated_cv", repeated_cv)

}
