% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task.R
\name{create_task}
\alias{create_task}
\title{Function that given a dataframe and variable creates an mlr3 task}
\usage{
create_task(df, target_var)
}
\arguments{
\item{df}{input data frame}

\item{target_var}{target variable}
}
\value{
returns a mlr3 task object
}
\description{
Function that given a dataframe and variable creates an mlr3 task
}
\examples{
create_task(iris, "Species")
create_task(mtcars, 1)

}
