% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_console.R, R/shiny_helpers.R
\name{debug_console}
\alias{debug_console}
\title{Function to print some information in the console.}
\usage{
debug_console(message, lvl = "INFO")

debug_console(message, lvl = "INFO")
}
\arguments{
\item{message}{message print to console}

\item{lvl}{level for the message}
}
\description{
Function to print some information in the console.

Function to print some logs into the console
}
