% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{generate_descr_report_cor}
\alias{generate_descr_report_cor}
\title{Function for creating correlation text part of descriptive report}
\usage{
generate_descr_report_cor(cor_matrix, feature, cor_thresholds = c(0.3, 0.5))
}
\arguments{
\item{cor_matrix}{correlation matrix}

\item{feature}{feature name}
}
\value{
returns string
}
\description{
Function for creating correlation text part of descriptive report
}
